Split runtime event normalizers
